v11 实现秒杀商品功能

+++ src/main/java/com/example/seckill/controller/GoodsController.java
@@ -128,7 +128,7 @@ public class GoodsController {
    * @return
    */
   @RequestMapping("/toDetail/{goodsId}")
-  public String toDetails(Model model, User user, @PathVariable Integer goodsId) {
+  public String toDetails(Model model, User user, @PathVariable Long goodsId) {
 
     //添加user用并且传输到前端
     model.addAttribute("user", user);
