+++ src/main/java/com/example/seckill/service/impl/GoodsServiceImpl.java
@@ -42,7 +42,7 @@ public class GoodsServiceImpl extends ServiceImpl<GoodsMapper, Goods> implements
    * @return
    */
   @Override
-  public GoodsVO findGoodsByGoodsId(Integer goodsId) {
+  public GoodsVO findGoodsByGoodsId(Long goodsId) {
 
     System.out.println("秒杀商品详情 = " + goodsMapper.findGoodsByGoodsId(goodsId));
     return goodsMapper.findGoodsByGoodsId(goodsId);
